+++ src/pool/index.ts
@@ -1,2 +1,3 @@
 export { PoolItem } from './pool_item.js';
 export { pool } from './pool.js';
+export { getDefaultPoolConcurrency } from './utilities.js';