+++ src/utilities/index.ts
@@ -3,7 +3,6 @@ export { createMutex, lockMutexSync, lockMutex, unlockMutex } from './mutex.js';
 export { createSharedUint32, getValue, setValue } from './shared_uint32.js';
 export { createCounter, getCount, incr, decr } from './counter.js';
 export { callsites } from './callsites.js';
-export { EventEmitter as TypedEmitter } from 'events';
 
 export type { Mutex } from './mutex.js';
 export type { SharedUint32 } from './shared_uint32.js';